Corrige code smells e refatora projeto para nome e namespace corretos.
