+++ IdentityStudie.Domain/Entities/BaseEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace Identity.Domain.Entities
+﻿namespace IdentityStudie.Domain.Entities
 {
     public abstract class BaseEntity
     {
