+++ IdentityStudie.Domain/Entities/Category.cs
@@ -1,4 +1,4 @@
-﻿namespace Identity.Domain.Entities
+﻿namespace IdentityStudie.Domain.Entities
 {
     public class Category : BaseEntity
     {
